pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "wiki-rag"
version = "0.1.0"
description = "Wikipedia RAG system"
readme = "README.md"
authors = [
    {name = "Roy Rinberg", email = "royrinberg@gmail.com"},
]
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "cryptography>=42.0.5",
    "faiss-cpu>=1.7.4",
    "langchain-community>=0.3.23",
    "sentence-transformers>=4.1.0",
    "wikipedia>=1.4.0",
    "pandas>=2.3.0",
]

[project.optional-dependencies]
dev = [
    "ipykernel>=6.29.5",
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
]

[project.scripts]
rag-server = "wiki_rag.rag_server:main"

[tool.setuptools.packages.find]
include = ["wiki_rag*"]

[tool.setuptools.package-data]
"*" = ["*.json", "*.txt", "*.md"]